pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Tang Ziya", email = "tcztzy@gmail.Com" }
]
dependencies = [
    "langchain-openai>=0.2.3",
    "langchain-chroma>=0.1.4",
    "autogen-agentchat[retrievechat]~=0.2",
    "langchain-text-splitters>=0.3.0",
    "langchain-community>=0.3.3",
    "langchain>=0.3.4",
    "pyautogen>=0.3.1",
    "swarmx>=0.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/agents"]
